Classes used to implement a Keras model
The numerai.py defines the Numerai object. This implements basic
functionality for the Numerai tournament.
The naikeras.py defines the NumeraiKeras object. This object
encapsulates Tensorflow/Keras model functionality. It includes
objects (classes) to plot the loss after each epoch of training and
another one to record the best epoch during training. Best is either
lowest validation loss (regression) or highest accuracy (for a classifier).
THe naiMLP.py is an example of a Keras model using NumeraiKeras which
uses Numerai as the base class. Note that this does not include
training. One simply assembles the model and saves it with no weights.
That empty model is opened and trained in the next step of the pipeline.