A bash script to run training of models.
The script parses configuration settings then invokes the Python program
train.py with those settings as command line options.
During training, model checkpoints are saved to a top level directory
having the same name as the model (e.g. ./mlp).
Note that an option is to use TensorBoard to monitor training
convergence or to use matplotlib to plot training and validation loss.
The latter can be used in a Jupyter Notebook or with ssh X11 Forwarding.