A set of scripts to automate predictions
The newround.sh script creates a new round directory and copies the
prediction related files into it.
The trigger.py Python program waits for the next round to open. The
most recent round number is persisted in the file 'last_round'. This
file is updated when a new round opens.
The tournament.py simply prints the new round number. It is deprecated
and will be removed.
The trigger.sh script calls trigger.py to wait for a round to open then
calls newround.sh to create a directory for the round predictions and
data. The live data is downloaded (download.py) and transformed (etl.py)
then the predict.sh script is called to run inference and upload
predictions.